+++ tests/bmds3/models/test_continuous.py
@@ -8,7 +8,7 @@ from bmds.bmds3.constants import BMDS_BLANK_VALUE, DistType, PriorClass
 from bmds.bmds3.models import continuous
 from bmds.bmds3.types.continuous import ContinuousModelSettings
 
-from .run3 import RunBmds3
+from ..run3 import RunBmds3
 
 
 class TestPriorOverrides:
