+++ tests/bmds3/types/test_common.py
@@ -35,3 +35,5 @@ class TestNumpyFloatArray:
         for data in ["a", ["a", "a"], None]:
             with pytest.raises(ValidationError):
                 ExampleModel(d=data)
+
+
